feat(registry): add ks, pearson, chi2 to registry
Allow for changing parameters of comparison functions
