+++ popmon/config.py
@@ -17,10 +17,9 @@
 # IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 from pathlib import Path
-from typing import Literal, Optional, Union
+from typing import Dict, List, Optional, Union
 
 from pydantic import BaseModel, BaseSettings
-from pydantic.fields import Field
 
 # Global configuration for the joblib parallelization. Could be used to change the number of jobs, and/or change
 # the backend from default (loki) to 'multiprocessing' or 'threading'.
@@ -53,9 +52,7 @@ class HistogramSectionModel(BaseModel):
     name = "Histograms"
     description = "Histograms of the last few time slots (default: 2)."
 
-    hist_names: list[
-        Literal["heatmap", "heatmap_column_normalized", "heatmap_row_normalized"]
-    ] = [
+    hist_names: List[str] = [
         "heatmap",
         "heatmap_column_normalized",
         "heatmap_row_normalized",
@@ -72,7 +69,14 @@ class HistogramSectionModel(BaseModel):
         "heatmap_column_normalized": "The column-normalized heatmap allows for comparing of time bins when the counts in each bin vary.",
         "heatmap_row_normalized": "The row-normalized heatmaps allows for monitoring one value over time.",
     }
+
+    """
+    plot_hist_n: plot histograms for last 'n' periods. default is 2 (optional)
+    """
     plot_hist_n: int = 2
+    """
+    cmap: colormap for histogram heatmaps
+    """
     cmap: str = "autumn_r"
 
 
@@ -102,12 +106,44 @@ class Section(BaseModel):
     traffic_lights: TrafficLightsSection = TrafficLightsSection()
 
 
-def get_stats():
-    from popmon.analysis.comparison.comparisons import Comparisons
+class Report(BaseModel):
+    """Report-specific configuration"""
 
-    comparisons = Comparisons.get_descriptions()
+    """
+    skip_empty_plots: if false, also show empty plots in report with only nans or zeroes (optional)
+    """
+    skip_empty_plots: bool = True
 
-    stats = [
+    """
+    last_n: plot statistic data for last 'n' periods (optional)
+    """
+    last_n: int = 0
+
+    """
+    skip_first_n: in plot skip first 'n' periods. last_n takes precedence (optional)
+    """
+    skip_first_n: int = 0
+
+    """
+    skip_last_n: in plot skip last 'n' periods. last_n takes precedence (optional)
+    """
+    skip_last_n: int = 0
+
+    """
+    report_filepath: the file path where to output the report (optional)
+    """
+    report_filepath: Optional[Union[str, Path]] = None
+
+    """
+    if True, show all the generated statistics in the report (optional)
+    if set to False, then smaller show_stats (see below)
+    """
+    extended_report: bool = True
+
+    """
+    show_stats: list of statistic name patterns to show in the report. If None, show all (optional)
+    """
+    show_stats: List[str] = [
         "distinct*",
         "filled*",
         "nan*",
@@ -124,68 +160,86 @@ def get_stats():
         "*chi2_norm*",
         "*zscore*",
         "n_*",
+        "*jsd*",
+        "*psi*",
+        "*max_prob_diff*",
     ]
 
-    for key in comparisons.keys():
-        stats.append(f"*{key}*")
-
-    return stats
-
-
-class Report(BaseModel):
-    """Report-specific configuration"""
-
-    skip_empty_plots: bool = True
-    last_n: int = 0
-    skip_first_n: int = 0
-    skip_last_n: int = 0
-    report_filepath: Optional[Union[str, Path]] = None
-    # if set to false, then smaller show_stats
-    # if limited report is selected, check if stats list is provided, if not, get a default minimal list
-    # show_stats = show_stats if not extended_report else None
-    extended_report: bool = True
-    show_stats: list[str] = Field(default_factory=get_stats)
-    section: Section = Section()
+    """
+    top_n: limit of number of categorical items to plot (default: 20)
+    """
     top_n: int = 20
+
+    section: Section = Section()
 
 
 class Comparison(BaseModel):
-    window = 10
-    shift = 1
+    """
+    window: size of rolling window and/or trend detection. default is 10.
+    """
+
+    window: int = 10
+    """
+    shift: shift of time-bins in rolling/expanding window. default is 1.
+    """
+    shift: int = 1
 
 
 class Monitoring(BaseModel):
-    monitoring_rules: dict[str, list[float]] = {
+    """
+    monitoring_rules: monitoring rules to generate traffic light alerts.
+    The default setting is:
+
+    .. code-block:: python
+
+        monitoring_rules = {
+            "*_pull": [7, 4, -4, -7],
+            "*_zscore": [7, 4, -4, -7],
+            "[!p]*_unknown_labels": [0.5, 0.5, 0, 0],
+        }
+
+    Note that the (filename based) wildcards such as * apply to all statistic names matching that pattern.
+    For example, ``"*_pull"`` applies for all features to all statistics ending on "_pull".
+    You can also specify rules for specific features and/or statistics by leaving out wildcard and putting the
+    feature name in front. E.g.
+
+    .. code-block:: python
+
+        monitoring_rules = {
+            "featureA:*_pull": [5, 3, -3, -5],
+            "featureA:nan": [4, 1, 0, 0],
+            "*_pull": [7, 4, -4, -7],
+            "nan": [8, 1, 0, 0],
+        }
+
+    In case of multiple rules could apply for a feature's statistic, the most specific one applies.
+    So in case of the statistic "nan": "featureA:nan" is used for "featureA", and the other "nan" rule
+    for all other features.
+    """
+
+    monitoring_rules: Dict[str, List[Union[float, int]]] = {
         "*_pull": [7, 4, -4, -7],
         "*_zscore": [7, 4, -4, -7],
         "[!p]*_unknown_labels": [0.5, 0.5, 0, 0],
     }
-    pull_rules: dict[str, list[float]] = {"*_pull": [7, 4, -4, -7]}
+
+    """
+    pull_rules: red and yellow (possibly dynamic) boundaries shown in plots in the report.
+    Default is:
+
+    .. code-block:: python
+
+        pull_rules = {"*_pull": [7, 4, -4, -7]}
+
+    This means that the shown yellow boundaries are at -4, +4 standard deviations around the (reference) mean,
+    and the shown red boundaries are at -7, +7 standard deviations around the (reference) mean.
+    Note that the (filename based) wildcards such as * apply to all statistic names matching that pattern.
+    (The same string logic applies as for monitoring_rules.)
+    """
+    pull_rules: Dict[str, List[Union[float, int]]] = {"*_pull": [7, 4, -4, -7]}
 
 
 class Settings(BaseSettings):
     report: Report = Report()
     comparison: Comparison = Comparison()
     monitoring: Monitoring = Monitoring()
-
-    @classmethod
-    def get_keys(cls):
-        aliases = {}
-        ambiguous = []
-        for key, value in cls.schema()["properties"].items():
-            if key in aliases:
-                ambiguous.append(key)
-                del aliases[key]
-            elif key in ambiguous:
-                continue
-
-            if "allOf" in value:
-                for skey, svalue in value["default"].items():
-                    if skey in aliases:
-                        ambiguous.append(key)
-                        del aliases[key]
-                    else:
-                        aliases[skey] = (key, skey)
-            else:
-                aliases[key] = key
-        return aliases
