+++ popmon/stats/numpy.py
@@ -18,68 +18,8 @@
 # CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 
-import warnings
-
 import numpy as np
 import pandas as pd
-from scipy import stats
-
-from popmon.analysis.comparison.comparisons import Comparisons
-
-
-def fraction_of_true(bin_labels, bin_entries):
-    """Compute fraction of 'true' labels
-
-    :param bin_labels: Array containing numbers whose mean is desired. If `a` is not an
-        array, a conversion is attempted.
-    :param bin_entries: Array containing weights for the elements of `a`. If `weights` is not an
-        array, a conversion is attempted.
-    :return: fraction of 'true' labels
-    """
-    bin_labels = np.array(bin_labels)
-    bin_entries = np.array(bin_entries)
-    assert len(bin_labels) == len(bin_entries)
-
-    def replace(bl):
-        if bl in {"True", "true"}:
-            return True
-        elif bl in {"False", "false"}:
-            return False
-        return np.nan
-
-    # basic checks: dealing with boolean labels
-    # also accept strings of 'True' and 'False'
-    if len(bin_labels) == 0 or len(bin_labels) > 4 or np.sum(bin_entries) == 0:
-        return np.nan
-    if not np.all([isinstance(bl, (bool, np.bool_)) for bl in bin_labels]):
-        if not np.all(
-            [isinstance(bl, (str, np.str_, np.string_)) for bl in bin_labels]
-        ):
-            return np.nan
-        # all strings from hereon
-        n_true = (bin_labels == "True").sum() + (bin_labels == "true").sum()
-        n_false = (bin_labels == "False").sum() + (bin_labels == "false").sum()
-        n_nan = (
-            (bin_labels == "NaN").sum()
-            + (bin_labels == "nan").sum()
-            + (bin_labels == "None").sum()
-            + (bin_labels == "none").sum()
-            + (bin_labels == "Null").sum()
-            + (bin_labels == "null").sum()
-        )
-        if n_true + n_false + n_nan != len(bin_labels):
-            return np.nan
-        # convert string to boolean
-        bin_labels = np.array([replace(bl) for bl in bin_labels])
-
-    sum_true = np.sum([be for bl, be in zip(bin_labels, bin_entries) if bl])
-    sum_false = np.sum([be for bl, be in zip(bin_labels, bin_entries) if not bl])
-    sum_entries = sum_true + sum_false
-    if sum_entries == 0:
-        # all nans scenario
-        return np.nan
-    # exclude nans from fraction
-    return (1.0 * sum_true) / sum_entries
 
 
 def mean(a, weights=None, axis=None, dtype=None, keepdims=False, ddof=0):
@@ -215,179 +155,6 @@ def quantile(a, q, weights=None, axis=None, keepdims: bool = False):
 
         y = y.reshape(shape)
         return y
-
-
-def _not_finite_to_zero(x):
-    res = x.copy()
-    res[~np.isfinite(res)] = 0
-    return res
-
-
-def uu_chi2(n, m, verbose=False):
-    """Normalized Chi^2 formula for two histograms with different number of entries
-
-    Copyright ROOT:
-    Formulas translated from c++ to python, but formulas otherwise not modified.
-    Reference: https://root.cern.ch/doc/master/classTH1.html#a6c281eebc0c0a848e7a0d620425090a5
-    GNU License: https://root.cern.ch/license
-    All modifications copyright ING WBAA.
-
-    :param n: 1d array with bin counts of the reference set
-    :param m: 1d array with bin counts of the test set
-    :param bool verbose: if true, print warnings in case of empty histograms
-    :return: tuple of floats (chi2_value, chi2_norm, z_score, p_value, res)
-    """
-    if len(n) == 0 or len(m) == 0:
-        raise ValueError("Input histogram(s) has zero size.")
-    if len(n) != len(m):
-        raise ValueError("Input histograms have unequal size.")
-
-    N = np.sum(n)
-    M = np.sum(m)
-
-    if N == 0 or M == 0:
-        if verbose:
-            warnings.warn(
-                "Input histogram(s) is empty and cannot be renormalized. Chi2 is undefined."
-            )
-        return np.nan, np.nan, np.nan, np.nan, [0] * len(n)
-
-    # remove all zero entries in the sum, to present division by zero for individual bins
-    z = n + m
-    n = n[z != 0]
-    m = m[z != 0]
-
-    dof = ((n != 0) | (m != 0)).sum() - 1
-    chi2_value = _not_finite_to_zero(((M * n - N * m) ** 2) / (n + m)).sum() / M / N
-
-    chi2_norm = chi2_value / dof if dof > 0 else np.nan
-    p_value = stats.chi2.sf(chi2_value, dof)
-    z_score = -stats.norm.ppf(p_value)
-
-    p = (n + m) / (N + M)
-
-    if (p == 1).any():
-        # unusual case of (only) one bin with p==1, avoids division with zero below
-        res = np.array([np.nan] * len(p))
-    else:
-        res = _not_finite_to_zero(
-            (n - N * p) / np.sqrt(N * p) / np.sqrt((1 - N / (N + M)) * (1 - p))
-        )
-
-    return chi2_value, chi2_norm, z_score, p_value, res
-
-
-def ks_test(hist_1, hist_2):
-    """KS-test for two histograms with different number of entries
-
-    Copyright ROOT:
-    Formulas translated from c++ to python, but formulas otherwise not modified.
-    Reference: link: https://root.cern.ch/doc/master/classTH1.html#TH1:KolmogorovTest
-    GNU license: https://root.cern.ch/license
-    All modifications copyright ING WBAA.
-
-    :param hist_1: 1D array with bin counts of the histogram_1
-    :param hist_2: 1D array with bin counts of the histogram_2
-
-    :return: ks_score: Kolmogorov-Smirnov Test score
-    :rtype: float
-    """
-    if len(hist_1) == 0 or len(hist_2) == 0:
-        raise ValueError("Input histogram(s) has zero size.")
-    if len(hist_1) != len(hist_2):
-        raise ValueError("Input histograms have unequal size.")
-
-    sum_1 = np.sum(hist_1)
-    sum_2 = np.sum(hist_2)
-    if sum_1 == 0 or sum_2 == 0:
-        return np.nan
-
-    normalized_cumsum_1 = np.cumsum(hist_1) / sum_1
-    normalized_cumsum_2 = np.cumsum(hist_2) / sum_2
-
-    d = np.abs(normalized_cumsum_1 - normalized_cumsum_2)
-
-    return np.max(d) * np.sqrt(sum_1 * sum_2 / (sum_1 + sum_2))
-
-
-def ks_prob(testscore):
-    """KS-probability corresponding ti KS test score
-
-    Copyright ROOT:
-    Formulas translated from c++ to python, but formulas otherwise not modified.
-    Reference: https://root.cern.ch/doc/master/classTH1.html#TH1:KolmogorovTest
-    GNU license: https://root.cern.ch/license
-    All modifications copyright ING WBAA.
-
-    :param float testscore: Kolmogorov-Smirnov test score
-
-    :return: approximate pvalue for the Kolmogorov-Smirnov test score
-    :rtype: float
-    """
-    fj = np.array([-2, -8, -18, -32])
-    r = np.zeros(4)
-
-    w = 2.50662827
-    c = np.array([-1.2337005501361697, -11.103304951225528, -30.842513753404244])
-
-    u = abs(testscore)
-    pvalue = np.nan
-    if u < 0.2:
-        pvalue = 1
-    elif u < 0.755:
-        v = np.power(u, -2)
-        pvalue = 1 - w * np.exp(c * v).sum() / u
-    elif u < 6.8116:
-        v = np.power(u, 2)
-        max_j = int(max(1, round(3.0 / u)))
-        r[:max_j] = np.exp(fj[:max_j] * v)
-        pvalue = 2 * (r[0] - r[1] + r[2] - r[3])
-
-    return pvalue
-
-
-@Comparisons.register(
-    key="max_prob_diff",
-    description="The largest absolute difference between all bin pairs of two normalized histograms (one histogram in a time slot and one in {ref})",
-)
-def googl_test(bins_1, bins_2):
-    """Google-paper test
-
-    Reference link: https://mlsys.org/Conferences/2019/doc/2019/167.pdf
-
-    :param bins_1: first array of bin entries
-    :param bins_2: second array of entries
-
-    :return: maximum difference between the two entry distributions
-    :rtype: float
-    """
-
-    def dist(bins):
-        sum_ = np.sum(bins)
-        return bins / sum_ if sum_ else bins
-
-    return np.max(np.abs(dist(bins_1) - dist(bins_2)))
-
-
-@Comparisons.register(key="psi", description="Population Stability Index")
-def population_stability_index(p, q):
-    epsilon = 10e-6
-    p += epsilon
-    q += epsilon
-    return np.sum((p - q) * np.log(p / q))
-
-
-def kullback_leibler_divergence(p, q):
-    epsilon = 10e-6
-    p += epsilon
-    q += epsilon
-    return np.sum(p * np.log(p / q))
-
-
-@Comparisons.register(key="jsd", description="Jensen-Shannon Divergence")
-def jensen_shannon_divergence(p, q):
-    m = 0.5 * (p + q)
-    return 0.5 * (kullback_leibler_divergence(p, m) + kullback_leibler_divergence(q, m))
 
 
 def probability_distribution_mean_covariance(entries_list):
